client side user parts refactored
