--- client/src/components/UI/ContactinfoModal.css
@@ -1,31 +1,0 @@
-.contactInfoModal__wrapper {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.8);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.contactInfoModal__content {
-  background: white;
-  border-radius: 8px;
-  display: flex;
-  flex-direction: column;
-  width: 50%;
-  height: 70%;
-  align-items: center;
-  justify-content: center;
-}
-
-.contactInfoModal__form {
-  height: 30%;
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-}
