+++ client/src/components/modals/AddQuestionModal.js
@@ -1,97 +1,80 @@
 import React, { useState, Fragment, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import AddedMedia from "./AddedMedia";
-import camelToSentence from "../../utils/camelToSentence";
+import AddedMedia from "../UI/AddedMedia";
 import * as quizActions from "../../store/actions/quizActions";
-import { makeStyles } from "@material-ui/core/styles";
-import { Paper, Grid, Typography, Button, TextField } from "@material-ui/core";
-//STYLES
-const useStyles = makeStyles((theme) => ({
-	root: {
-		flexGrow: 1
-	},
-	paper: {
-		padding: theme.spacing(1),
-		textAlign: "center",
-		color: theme.palette.text.secondary,
-		width: "50%"
-	}
-}));
 
-//these two lists should probably be passed in to this component via props from the main QuizMaker component
-const supportedFileTypes = [
-	//image
-	"image/jpeg",
-	"image/png",
-	"image/svg",
-	//video
-	"video/mp4",
-	"video/webm",
-	"video/ogg",
-	//audio
-	"audio/mpeg",
-	"audio/ogg",
-	"audio/wav"
-];
+import { modalRootStyles, addQuestionModalStyles } from "../../style/modalStyles";
+import camelToSentence from "../../utils/camelToSentence";
+import supportedFileTypes from "../../utils/supportedFileTypes";
+import { Paper, Grid, Typography, Button, TextField } from "@material-ui/core";
 
 const AddQuestionModal = ({ closeModal, quiz, questionToEdit }) => {
-	const classes = useStyles();
+	const rootClasses = modalRootStyles();
+	const classes = addQuestionModalStyles();
 	const dispatch = useDispatch();
 
-	const [addedMedia, setAddedMedia] = useState([]);
-	const [retrivedMedia, setRetrivedMedia] = useState([])
-	const [questionType, setQuestionType] = useState("multipleChoice");
-	const [multipleChoiceOptions, setMultipleChoiceOptions] = useState(["", ""]);
-	const [selectedMultipleChoiceOption, setSelectedMultipleChoiceOption] = useState(null);
-	const [selectedTrueFalse, setSelectedTrueFalse] = useState();
-	const [question, setQuestion] = useState("");
-	const [points, setPoints] = useState(null);
-	const questionTypes = ["trueFalse", "multipleChoice"];
+	const [ addedMedia, setAddedMedia ] = useState([]);
+	const [ retrivedMedia, setRetrivedMedia ] = useState([]);
+	const [ questionType, setQuestionType ] = useState("multipleChoice");
+	const [ multipleChoiceOptions, setMultipleChoiceOptions ] = useState([ "", "" ]);
+	const [ selectedMultipleChoiceOption, setSelectedMultipleChoiceOption ] = useState(null);
+	const [ selectedTrueFalse, setSelectedTrueFalse ] = useState();
+	const [ question, setQuestion ] = useState("");
+	const [ points, setPoints ] = useState(null);
+	const questionTypes = [ "trueFalse", "multipleChoice" ];
 	const qToEdit = useSelector((state) =>
 		state.quiz.quizQuestions.find((question) => {
 			return question._id === questionToEdit;
 		})
 	);
 
+	useEffect(
+		() => {
+			if (qToEdit) {
+				setQuestion(qToEdit.question);
+				setQuestionType(qToEdit.questionType);
+				setMultipleChoiceOptions(
+					qToEdit.answers.multipleChoiceOptions ? qToEdit.answers.multipleChoiceOptions : [ "", "" ]
+				);
+				setSelectedMultipleChoiceOption(qToEdit.answers.multipleChoiceAnswer);
+				setSelectedTrueFalse(qToEdit.answers.trueFalseAnswer);
+				setPoints(qToEdit.points);
+				qToEdit.media.map(async (media) => {
+					if (media.mediaType === "text/plain") {
+						setRetrivedMedia(
+							retrivedMedia.push({ file: { mediaType: "text/plain", text: media.data }, id: Date.now() })
+						);
+					} else {
+						await urltoFile(
+							`data:${media.mediaType};base64,${media.data}`,
+							`${media.name}`,
+							`${media.mediaType}`
+						).then((file) => {
+							setRetrivedMedia(retrivedMedia.push({ file: file, id: Date.now() }));
+						});
+					}
+				});
+				setAddedMedia(retrivedMedia);
+				console.log(retrivedMedia);
+			}
+		},
+		[ qToEdit ]
+	);
+
+	//FUNCTIONS 
 	const urltoFile = async (url, filename, mimeType) => {
 		return await fetch(url)
-			.then(function (res) {
+			.then(function(res) {
 				return res.arrayBuffer();
 			})
-			.then(function (buf) {
-				return new File([buf], filename, { type: mimeType });
+			.then(function(buf) {
+				return new File([ buf ], filename, { type: mimeType });
 			});
-	}
+	};
 	//Usage example:
 	// urltoFile("data:text/plain;base64,aGVsbG8gd29ybGQ=", "hello.txt", "text/plain").then(function(file) {
 	// 	console.log(file);
 	// });
-
-	
-	useEffect(() => {
-		if (qToEdit) {
-			setQuestion(qToEdit.question);
-			setQuestionType(qToEdit.questionType);
-			setMultipleChoiceOptions(
-				qToEdit.answers.multipleChoiceOptions ? qToEdit.answers.multipleChoiceOptions : ["", ""]
-			);
-			setSelectedMultipleChoiceOption(qToEdit.answers.multipleChoiceAnswer);
-			setSelectedTrueFalse(qToEdit.answers.trueFalseAnswer);
-			setPoints(qToEdit.points);
-			qToEdit.media.map(async (media) => {
-				if (media.mediaType === "text/plain") {
-					setRetrivedMedia(retrivedMedia.push( { file: { mediaType: "text/plain", text: media.data }, id: Date.now() }))
-				} else {
-					await urltoFile(`data:${media.mediaType};base64,${media.data}`, `${media.name}`, `${media.mediaType}`).then((file) => {
-					setRetrivedMedia(retrivedMedia.push({file: file, id: Date.now()}))
-				});
-				}
-				})
-			setAddedMedia(retrivedMedia)
-			console.log(retrivedMedia)
-			
-		}
-	}, [qToEdit]);
 
 	//HANDLERS
 	const handleAddText = () => {
@@ -114,7 +97,8 @@ const AddQuestionModal = ({ closeModal, quiz, questionToEdit }) => {
 		}
 		setAddedMedia([
 			...addedMedia,
-			{file: file,
+			{
+				file: file,
 				//need some kind of id to be able to remove media after adding it
 				id: Date.now()
 			}
@@ -172,7 +156,7 @@ const AddQuestionModal = ({ closeModal, quiz, questionToEdit }) => {
 			.filter((media) => media.file.mediaType === "text/plain" && media.file.data !== "")
 			.map((obj) => obj.file);
 		const questionObject = {
-			id: qToEdit? qToEdit._id : new Date().getUTCMilliseconds().toString(),
+			id: qToEdit ? qToEdit._id : new Date().getUTCMilliseconds().toString(),
 			questionType: questionType,
 			//adding text separately because FormData will ignore it
 			media: [ ...addedText ],
@@ -191,14 +175,15 @@ const AddQuestionModal = ({ closeModal, quiz, questionToEdit }) => {
 		addedMedia.forEach((media) => {
 			formData.append("file", media.file);
 		});
-		qToEdit? await dispatch(quizActions.editQuestion(formData)):
-		await dispatch(quizActions.addNewQuestion(formData));
+		qToEdit
+			? await dispatch(quizActions.editQuestion(formData))
+			: await dispatch(quizActions.addNewQuestion(formData));
 		closeModal();
 	};
 
 	//RETURN
 	return (
-		<Grid container spacing={2} style={{ width: "80%" }}>
+		<Grid className={rootClasses.root} container spacing={2}>
 			<Grid item xs={12} className={classes.paper}>
 				<Typography className={classes.paper} variant="h5">
 					Add Question
