--- client/src/dummy-data/contacts.js
@@ -1,104 +1,0 @@
-export const groups = [
-  {
-    name: "1A",
-    members: [
-      {
-        name: "Sarah Smith",
-        email: "sarah@gmail.com",
-      },
-      {
-        name: "John Brown",
-        email: "john@gmail.com",
-      },
-      {
-        name: "Greg Wallace",
-        email: "greg@gmail.com",
-      },
-    ],
-  },
-  {
-    name: "1B",
-    members: [
-      {
-        name: "Tim Rice",
-        email: "tim@gmail.com",
-      },
-      {
-        name: "Jennifer Baxter",
-        email: "jennifer@gmail.com",
-      },
-      {
-        name: "Mary O'Toole",
-        email: "mary@gmail.com",
-      },
-      {
-        name: "Aoife Flynn",
-        email: "aoife@gmail.com",
-      },
-    ],
-  },
-  {
-    name: "1C",
-    members: [
-      {
-        name: "Harry Hill",
-        email: "harry@gmail.com",
-      },
-      {
-        name: "Roger Allam",
-        email: "roger@gmail.com",
-      },
-    ],
-  },
-];
-
-export const contacts = [
-  {
-    name: "Harry Hill",
-    email: "harry@gmail.com",
-  },
-  {
-    name: "Roger Allam",
-    email: "roger@gmail.com",
-  },
-  {
-    name: "Tim Rice",
-    email: "tim@gmail.com",
-  },
-  {
-    name: "Jennifer Baxter",
-    email: "jennifer@gmail.com",
-  },
-  {
-    name: "Mary O'Toole",
-    email: "mary@gmail.com",
-  },
-  {
-    name: "Aoife Flynn",
-    email: "aoife@gmail.com",
-  },
-  {
-    name: "Sarah Smith",
-    email: "sarah@gmail.com",
-  },
-  {
-    name: "John Brown",
-    email: "john@gmail.com",
-  },
-  {
-    name: "Greg Wallace",
-    email: "greg@gmail.com",
-  },
-  {
-    name: "Sarah Smith",
-    email: "sarah@gmail.com",
-  },
-  {
-    name: "John Brown",
-    email: "john@gmail.com",
-  },
-  {
-    name: "Greg Wallace",
-    email: "greg@gmail.com",
-  },
-];
