+++ client/src/screens/authenticationScreens/Login.js
@@ -2,124 +2,93 @@ import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import * as authActions from "../../store/actions/authActions";
-
-import Avatar from "@material-ui/core/Avatar";
-import {
-  Button,
-  CssBaseline,
-  TextField,
-  Grid,
-  Typography,
-  Container,
-} from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
+import { Button, CssBaseline, TextField, Grid, Typography, Container, Avatar } from "@material-ui/core";
+import { useStyles } from "../../style/authScreensStyles";
 
 const Login = (props) => {
-  const [email, setEmail] = useState("testuser@test.com");
-  const [password, setPassword] = useState("12345678");
-  const dispatch = useDispatch();
-  const { from } = props.location.state || { from: { pathname: "/dashboard" } };
+	const dispatch = useDispatch();
+	const classes = useStyles();
 
-  const loginHandler = async () => {
-    const loginData = {
-      email,
-      password,
-    };
-    await dispatch(authActions.login(loginData));
-    props.history.push(from);
-  };
+	const [ email, setEmail ] = useState("testuser@test.com");
+	const [ password, setPassword ] = useState("12345678");
+	const { from } = props.location.state || { from: { pathname: "/dashboard" } };
 
-  const classes = useStyles();
+	//HANDLERS
+	const loginHandler = async () => {
+		const loginData = { email, password };
+		await dispatch(authActions.login(loginData));
+		props.history.push(from);
+	};
 
-  return (
-    <Container component="main" maxWidth="xs">
-      <CssBaseline />
-      <div className={classes.paper}>
-        <Avatar className={classes.avatar} />
-        <Typography component="h1" variant="h5">
-          SignIn
-        </Typography>
-        <form className={classes.form} noValidate>
-          <Grid container spacing={2}>
-            <Grid item xs={12}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                autoComplete="email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                minLength="8"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-              />
-            </Grid>
-          </Grid>
-          <Button
-            fullWidth
-            variant="contained"
-            color="primary"
-            className={classes.submit}
-            onClick={loginHandler}
-          >
-            Sign In
-          </Button>
-          <Grid container justify="flex-end">
-            <Grid item>
-              <Link to="/register">
-                <Typography href="#" variant="body2">
-                  Don"t have an account? Sign up
-                </Typography>
-              </Link>
-            </Grid>
-          </Grid>
-          <Grid container justify="flex-end">
-            <Grid item>
-              <Link to="/forgotPassword">
-                <Typography href="#" variant="body2">
-                  I forgot my password
-                </Typography>
-              </Link>
-            </Grid>
-          </Grid>
-        </form>
-      </div>
-    </Container>
-  );
+	//MAIN
+	return (
+		<Container component="main" maxWidth="xs">
+			<CssBaseline />
+			<div className={classes.paper}>
+				<Avatar className={classes.avatar} />
+				<Typography component="h1" variant="h5">
+					SignIn
+				</Typography>
+				<form className={classes.form} noValidate>
+					<Grid container spacing={2}>
+						<Grid item xs={12}>
+							<TextField
+								variant="outlined"
+								required
+								fullWidth
+								id="email"
+								label="Email Address"
+								name="email"
+								autoComplete="email"
+								value={email}
+								onChange={(e) => setEmail(e.target.value)}
+							/>
+						</Grid>
+						<Grid item xs={12}>
+							<TextField
+								variant="outlined"
+								required
+								fullWidth
+								name="password"
+								label="Password"
+								type="password"
+								id="password"
+								minLength="8"
+								value={password}
+								onChange={(e) => setPassword(e.target.value)}
+							/>
+						</Grid>
+					</Grid>
+					<Button
+						fullWidth
+						variant="contained"
+						color="primary"
+						className={classes.submit}
+						onClick={loginHandler}>
+						Sign In
+					</Button>
+					<Grid container justify="flex-end">
+						<Grid item>
+							<Link to="/register">
+								<Typography href="#" variant="body2">
+									Don"t have an account? Sign up
+								</Typography>
+							</Link>
+						</Grid>
+					</Grid>
+					<Grid container justify="flex-end">
+						<Grid item>
+							<Link to="/forgotPassword">
+								<Typography href="#" variant="body2">
+									I forgot my password
+								</Typography>
+							</Link>
+						</Grid>
+					</Grid>
+				</form>
+			</div>
+		</Container>
+	);
 };
-
-const useStyles = makeStyles((theme) => ({
-  paper: {
-    marginTop: theme.spacing(8),
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main,
-  },
-  form: {
-    width: "100%", // Fix IE 11 issue.
-    marginTop: theme.spacing(1),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-}));
 
 export default Login;
