--- client/src/screens/userScreens/Contacts.css
@@ -1,41 +1,0 @@
-.contacts__wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  height: 100%;
-}
-
-.contacts__title {
-  text-align: center;
-}
-
-.contacts__content {
-  width: 80%;
-  height: 80%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-direction: column;
-  overflow-y: auto;
-}
-
-.contacts__contacts {
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-}
-
-.contacts__contact {
-  flex-basis: 50%;
-  font-size: 1.5rem;
-  display: flex;
-  background-color: #f4f4f4;
-  border: 1px solid #333;
-  justify-content: space-between;
-  cursor: pointer;
-}
-
-.contacts__contact:hover {
-  background-color: #e7dede;
-}
