+++ client/src/style/authScreensStyles.js
@@ -1,0 +1,26 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+export const useStyles = makeStyles((theme) => ({
+      paper: {
+    marginTop: theme.spacing(8),
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+  },
+	form: {
+		width: "100%", // Fix IE 11 issue.
+		marginTop: theme.spacing(1)
+	},
+	submit: {
+		margin: theme.spacing(3, 0, 2)
+    },
+    actionButton: {
+        margin: theme.spacing(2 ,0, 1)
+    },
+	large: {
+		width: theme.spacing(12),
+		height: theme.spacing(12)
+	}
+}));
+
+
