+++ client/src/utils/supportedFileTypes.js
@@ -1,0 +1,16 @@
+ const supportedFileTypes = [
+	//image
+	"image/jpeg",
+	"image/png",
+	"image/svg",
+	//video
+	"video/mp4",
+	"video/webm",
+	"video/ogg",
+	//audio
+	"audio/mpeg",
+	"audio/ogg",
+	"audio/wav"
+];
+
+export default supportedFileTypes
